# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tunescape backend API.",
        "title": "Tunescape API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "List friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_friends.FriendProfile"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/friends/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "List incoming friend requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_friends.IncomingRequest"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Send a friend request",
                "parameters": [
                    {
                        "description": "receiver",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_friends.sendRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/friends/requests/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Accept a friend request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/friends/requests/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Reject a friend request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/friends/{friend_id}/compare": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns averaged audio feature scores and a 0–100 compatibility score. Returns 422 if the friend has no listening data yet.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Compare music tastes with a friend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Friend's user ID",
                        "name": "friend_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_friends.TasteComparison"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
        "/friends/{friend_id}/playlists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns playlists created via Tunescape. Returns 422 if the friend has not connected Spotify yet.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "Get a friend's playlists",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Friend's user ID",
                        "name": "friend_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gitlab_com_Uranury_tunescape_internal_playlist.Playlist"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gitlab_com_Uranury_tunescape_internal_playlist.Playlist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "embed_url": {
                    "type": "string"
                },
                "external_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "gitlab_com_Uranury_tunescape_pkg_apperrors.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_friends.FriendProfile": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "spotify_connected": {
                    "type": "boolean"
                },
                "spotify_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_friends.IncomingRequest": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "request_id": {
                    "type": "integer"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        },
        "internal_friends.TasteComparison": {
            "type": "object",
            "properties": {
                "compatibility_score": {
                    "type": "number"
                },
                "mine": {
                    "$ref": "#/definitions/internal_friends.TasteScores"
                },
                "theirs": {
                    "$ref": "#/definitions/internal_friends.TasteScores"
                }
            }
        },
        "internal_friends.TasteScores": {
            "type": "object",
            "properties": {
                "acousticness": {
                    "type": "number"
                },
                "danceability": {
                    "type": "number"
                },
                "energy": {
                    "type": "number"
                },
                "valence": {
                    "type": "number"
                }
            }
        },
        "internal_friends.sendRequestBody": {
            "type": "object",
            "required": [
                "receiver_id"
            ],
            "properties": {
                "receiver_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}